+++ packages/vidstack/src/player/core/api/player-props.ts
@@ -55,6 +55,10 @@ export const mediaPlayerProps: PropDeclarations<PlayerProps> = {
     attribute: false,
   }),
   frameRate: 30,
+  smallBreakpointX: 600,
+  largeBreakpointX: 980,
+  smallBreakpointY: 380,
+  largeBreakpointY: 600,
 };
 
 export interface MediaStateAccessors
@@ -196,4 +200,20 @@ export interface PlayerProps
    * @example `24`
    */
   frameRate: number;
+  /**
+   * The maximum player width in pixels for when the small breakpoint should be active.
+   */
+  smallBreakpointX: number;
+  /**
+   * The minimum player width in pixels for when the large breakpoint should begin being active.
+   */
+  largeBreakpointX: number;
+  /**
+   * The maximum player height in pixels for when the small breakpoint should be active.
+   */
+  smallBreakpointY: number;
+  /**
+   * The minimum player height in pixels for when the large breakpoint should begin being active.
+   */
+  largeBreakpointY: number;
 }