+++ packages/vidstack/src/player/core/outlet/tracks.ts
@@ -13,19 +13,19 @@ export class Tracks {
   private _onTracksChange() {
     const newTracks = [...this._media.$props.textTracks(), ...this._domTracks()];
 
+    for (const oldTrack of this._prevTracks) {
+      if (!newTracks.some((t) => t.id === oldTrack.id)) {
+        const track = oldTrack.id && this._media.textTracks.getById(oldTrack.id);
+        if (track) this._media.textTracks.remove(track);
+      }
+    }
+
     for (const newTrack of newTracks) {
       const id = newTrack.id || TextTrack.createId(newTrack);
       if (!this._media.textTracks.getById(id)) {
         // @ts-expect-error - override readonly
         newTrack.id = id;
         this._media.textTracks.add(newTrack);
-      }
-    }
-
-    for (const oldTrack of this._prevTracks) {
-      if (!newTracks.some((t) => t.id === oldTrack.id)) {
-        const track = oldTrack.id && this._media.textTracks.getById(oldTrack.id);
-        if (track) this._media.textTracks.remove(track);
       }
     }