+++ packages/vidstack/src/player/ui/menu/menu.tsx
@@ -28,7 +28,7 @@ import {
 import { $ariaBool } from '../../../utils/aria';
 import { isElementParent, onPress, setAttributeIfEmpty } from '../../../utils/dom';
 import { useMedia, type MediaContext } from '../../core/api/context';
-import { menuContext, type MenuContext } from './menu-context';
+import { menuContext, type MenuContext, type MenuObserver } from './menu-context';
 import { MenuFocusController } from './menu-focus-controller';
 
 declare global {
@@ -75,6 +75,7 @@ export class Menu extends Component<MenuAPI> {
 
   protected _menuButton: HTMLElement | null = null;
   protected _menuItems: HTMLElement | null = null;
+  protected _menuObserver: MenuObserver | null = null;
 
   protected _focus: MenuFocusController;
 
@@ -99,6 +100,7 @@ export class Menu extends Component<MenuAPI> {
       _disable: this._disable.bind(this),
       _attachMenuButton: this._attachMenuButton.bind(this),
       _attachMenuItems: this._attachMenuItems.bind(this),
+      _attachObserver: this._attachObserver.bind(this),
       _disableMenuButton: this._disableMenuButton.bind(this),
       _addSubmenu: this._addSubmenu.bind(this),
     });
@@ -130,6 +132,7 @@ export class Menu extends Component<MenuAPI> {
     this._removePopupMenu();
     this._menuButton = null;
     this._menuItems = null;
+    this._menuObserver = null;
   }
 
   protected _removePopupMenu() {
@@ -159,6 +162,14 @@ export class Menu extends Component<MenuAPI> {
       if (this.el.contains?.(this._menuItems)) {
         const menu = this.el!.cloneNode();
         menu.appendChild(this._menuItems);
+
+        requestAnimationFrame(() => {
+          if (!this.el) return;
+          const mediaRing = '--media-focus-ring',
+            mediaRingValue = getComputedStyle(this.el).getPropertyValue(mediaRing);
+          if (mediaRingValue) setStyle(menu as HTMLElement, mediaRing, mediaRingValue);
+        });
+
         scoped(() => {
           document.body.append(menu);
         }, this._media.scope);
@@ -230,6 +241,10 @@ export class Menu extends Component<MenuAPI> {
     this._updateMenuItemsHidden(peek(this._expanded));
   }
 
+  protected _attachObserver(observer: MenuObserver) {
+    this._menuObserver = observer;
+  }
+
   protected _updateMenuItemsHidden(expanded: boolean) {
     if (this._menuItems) setAttribute(this._menuItems, 'aria-hidden', ariaBool(!expanded));
   }
@@ -245,14 +260,31 @@ export class Menu extends Component<MenuAPI> {
 
     this._expanded.set((expanded) => !expanded);
     this._changeIdleTracking();
-
     tick();
-    if (isKeyboardEvent(event)) this._menuItems?.focus();
+
+    if (isKeyboardEvent(event)) {
+      this._menuItems?.focus();
+    }
+
     this._onChange(event);
   }
 
   protected _onChange(trigger?: Event) {
-    this.dispatch(peek(this._expanded) ? 'open' : 'close', { trigger });
+    const expanded = peek(this._expanded);
+    this.dispatch(expanded ? 'open' : 'close', { trigger });
+
+    if (!this._parentMenu) {
+      if (expanded) {
+        this._media.activeMenu?.close(trigger);
+        this._media.activeMenu = this;
+      } else {
+        for (const el of this._submenus) el.close(trigger);
+        this._media.activeMenu = null;
+      }
+    }
+
+    if (expanded) this._menuObserver?._onOpen?.(trigger);
+    else this._menuObserver?._onClose?.(trigger);
   }
 
   protected _isExpanded() {
@@ -353,10 +385,17 @@ export class Menu extends Component<MenuAPI> {
    */
   @method
   open(trigger?: Event) {
+    if (peek(this._expanded)) return;
+
     this._expanded.set(true);
     tick();
+
     this._onChange(trigger);
-    if (isKeyboardEvent(trigger)) this._menuItems?.focus();
+
+    if (isKeyboardEvent(trigger)) {
+      this._menuItems?.focus();
+    }
+
     this._changeIdleTracking(trigger);
   }
 
@@ -366,6 +405,8 @@ export class Menu extends Component<MenuAPI> {
    */
   @method
   close(trigger?: Event) {
+    if (!peek(this._expanded)) return;
+
     this._expanded.set(false);
     tick();